fastapi/requirements.txt
fastapi>=0.95.2
uvicorn>=0.22.0
mysql-connector-python>=8.0.33
pydantic>=1.10.8
PyMySQL>=1.0.3
python-dotenv>=1.0.0
SQLAlchemy>=2.0.15